+++ source/Messaging.Api/IncomingMessages/IncomingMessageQueueListener.cs
@@ -18,6 +18,7 @@ using System.Threading.Tasks;
 using Messaging.Api.Configuration;
 using Messaging.Application.Configuration;
 using Messaging.Application.IncomingMessages;
+using Messaging.Application.IncomingMessages.RequestChangeOfSupplier;
 using Messaging.Infrastructure.Configuration.Serialization;
 using Microsoft.Azure.Functions.Worker;
 using Microsoft.Extensions.Logging;
